Add ClientHandler.java
Responsible for handling communication with a single Telnet client, reads input from the client and sends responses,
implements ANSI escape codes for formatting and detects if localecho is on and adjusts accordingly.
Main purpose for this class is that each Telnet client runs in a separate thread, and this class isolates client-specific logic.