Add addContact function
Allows the user to add a new friend to the users existing friend group contact list.
If this friend exists in this contact list, it is not necessary to add that friend again.
On the other hand, if the frind is not on the contact list, the system will add the new
friend to the user's contact list.